JSON Caching
Instead of declaring the projects on the Javascript file, a separate JSON file was created and each it's gotten with axios get request. In order not to have so many requests, these projects are stored in LocalStorage with an expiration date of 1 week. On each page reload the localstorage is verified, if it exists and the date not expired, then the array of objects is loaded without any request.